% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{abs.physical_quantity}
\alias{abs.physical_quantity}
\title{abs.physical_quantity}
\usage{
\method{abs}{physical_quantity}(x)
}
\arguments{
\item{x}{: physical quantity}
}
\value{
The absolute value of a physical quantity
}
\description{
Absolute value of a physical quantity
}
\examples{
a <- physical_quantity(-5.4)
a <- a * u_m
a <- abs(a)
a
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

