% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{as.numeric}
\alias{as.numeric}
\alias{as.numeric.physical_quantity}
\title{as.numeric}
\usage{
\\method{as.numeric}{physical_quantity}(x, ...)
}
\arguments{
\item{x}{: physical_quantity}

\item{...}{: Other parameters}
}
\value{
a numeric object
}
\description{
Function to coerce a physical_quantity into a numeric
}
\examples{
v1 <- 5.6*u_km
v1n <- as.numeric(v1)
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

