% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{as.physical_quantity}
\alias{as.physical_quantity}
\title{as.physical_quantity}
\usage{
as.physical_quantity(v)
}
\arguments{
\item{v}{: a vector or list with values}
}
\value{
A physical_quantity
}
\description{
Function to coerce an object into a physical_quantity
}
\examples{
a <- as.physical_quantity(0.5)
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

