% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{data}
\name{u_meter}
\alias{u_ampere}
\alias{u_candela}
\alias{u_kelvin}
\alias{u_kilogram}
\alias{u_meter}
\alias{u_mole}
\alias{u_radian}
\alias{u_second}
\title{u_meter}
\format{An object of class \code{physical_quantity} of length 2.}
\usage{
u_meter

u_kilogram

u_second

u_kelvin

u_ampere

u_mole

u_candela

u_radian
}
\description{
1 meter
}
\keyword{datasets}

