% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{==.physical_quantity}
\alias{==.physical_quantity}
\title{==.physical_quantity}
\usage{
\method{==}{physical_quantity}(a, b)
}
\arguments{
\item{a, b}{: A physical_quantity objects}
}
\value{
Logical
}
\description{
Function to compare equality of physical_quantities
}
\examples{
a <- 3.4 * u_m
b <- 3.5 * u_m
a==b
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

