% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{get_unit_label}
\alias{get_unit_label}
\title{get_unit_label}
\usage{
get_unit_label(x)
}
\arguments{
\item{x}{: physical_quantity}
}
\value{
list with the unit label
}
\description{
Function to get the unit label
}
\examples{
v <- 4.56 * u_m/u_s
v.label <- get_unit_label(v)
v[1]
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

