% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{get_units.physical_quantity}
\alias{get_units.physical_quantity}
\title{get_units.physical_quantity}
\usage{
get_units.physical_quantity(q)
}
\arguments{
\item{q}{: physical_quantity}
}
\value{
character
}
\description{
Function to get  the units of a physical quantity
}
\examples{
l <- 1.5*u_m

}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

