% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{get_units}
\alias{get_units}
\title{get_units}
\usage{
get_units(q)
}
\arguments{
\item{q}{: A physical_quantity object}
}
\value{
A character vector
}
\description{
Function to get the units of a physical_quantity
}
\examples{
a <- 5.6*u_m
get_units(a)
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

