% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{get_value}
\alias{get_value}
\title{get_value}
\usage{
get_value(q)
}
\arguments{
\item{q}{: A physical_quantity object}
}
\value{
A numeric variable or vector
}
\description{
Function to get the value from a physical_quantity object
}
\examples{
length <- 1.5*u_m
get_value(length)
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

