% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{>.physical_quantity}
\alias{>.physical_quality}
\alias{>.physical_quantity}
\title{`>.physical_quantity`}
\usage{
`>.physical_quantity(a,b)`
}
\arguments{
\item{a, b}{: physical quantities}
}
\value{
A logical variable indicating if a is greater than b
}
\description{
Operator to check if a physical quantity is greater than another
of the same units
}
\examples{
a <- 5.4 * u_m
b <- 3.2 * u_m
c <- 8.9 * u_s
a > b
a > c
}

