% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{>=.physical_quantity}
\alias{>=.physical_quantity}
\title{>=.physical_quantity}
\usage{
\method{>=}{physical_quantity}(a, b)
}
\arguments{
\item{a, b}{: physical quantities or numeric values}
}
\value{
A logical variable specifying if a is greater than or equal to b
}
\description{
Operator to check if a physical quantity is greater than or equal to
another physical quantity of the same units or a numeric value
}
\examples{
side1 <- 1.5 * u_m
side2 <- 3.4 * u_m
side1 >= side2
side1 >= 3.4
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

