% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{is.na.physical_quantity}
\alias{is.na}
\alias{is.na.physical_quantity}
\title{is.na}
\usage{
\\method{is.na}{physical_quantity}(x)
}
\arguments{
\item{x}{: A physical_quantity object}
}
\value{
A logical value
}
\description{
A function to check if the value of a physical_quantity object is NA
}
\examples{
a <- 4.3*u_m
is.na(a)
}
\author{
Oscar Garcia-Cabrejo
}

