% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{is.physical_quantity}
\alias{is.physical_quantity}
\title{is.physical_quantity}
\usage{
is.physical_quantity(q)
}
\arguments{
\item{q}{: A physical_quantity object}
}
\value{
A logical value
}
\description{
A function to check if a given variable is
}
\examples{
a <- 5.6*u_m
is.physical_quantity(a)
}
\author{
Oscar Garcia-Cabrejo
}

