% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{is_equal}
\alias{!=.physical_quantity}
\alias{is_equal}
\title{is equal operator}
\usage{
\method{!=}{physical_quantity}(a, b)
}
\arguments{
\item{a, b}{: Physical_quantity objects to be compared}
}
\value{
Logical
}
\description{
Function to
}
\examples{
a <- 3.4 * u_m
b <- 5.6 * u_m
a!=b
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

