% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{is_nondimensional}
\alias{is_nondimensional}
\title{is_nondimensional}
\usage{
is_nondimensional(q)
}
\arguments{
\item{q}{: physical_quantity}
}
\value{
logical
}
\description{
Function to check if a physical quantity is nondimensional
}
\examples{
v1 <- physical_quantity(5.8)
is_nondimensional(v1)
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

