% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{<.physical_quantity}
\alias{<.physical_quantity}
\title{<.physical_quantity}
\usage{
\method{<}{physical_quantity}(a, b)
}
\arguments{
\item{a, b}{: numerical values and or physical quantities}
}
\value{
A logical variable indicating if a is less than b
}
\description{
Operator to  check if a physical quantity is less than another
of the same untis
}
\examples{
side1 <- 4.6 * u_m
side2 <- 3.4 * u_m
side2 < side1
side2 < 9.0
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

