% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{<=.physical_quantity}
\alias{<=.physical_quantity}
\title{<=.physical_quantity}
\usage{
\method{<=}{physical_quantity}(a, b)
}
\arguments{
\item{a, b}{: physical quantity or numerical value}
}
\value{
A logical value specifying if a <= b
}
\description{
Operator to check if a physical quantity is less or equal to another
physical quantity or numerical value
}
\examples{
side1 <- 3.4 * u_m
side2 <- 6.7 *u_m
side1 <= side2
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

