% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{minus.physical_quantity}
\alias{-}
\alias{-.physical_quantity}
\alias{minus.physical_quantity}
\title{`-.physical_quantity`}
\usage{
a-b
}
\arguments{
\item{a, b}{: these can be a numeric object and/or a physical_quantity object}
}
\value{
A physical_quantity
}
\description{
Define the substraction operation of physical_quantity class
}
\examples{
dist1 <- 5.7*u_m
dist2 <- 1.2*u_ft
dist3 <- dist1-dist2
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

