% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{data}
\name{non_si_units}
\alias{c_mass_solution}
\alias{c_mass_solvent}
\alias{c_volume_solution}
\alias{c_volume_solvent}
\alias{non_si_units}
\alias{u_Bq}
\alias{u_Ci}
\alias{u_Gy}
\alias{u_L}
\alias{u_L_per_s}
\alias{u_M}
\alias{u_Molar}
\alias{u_Pa_per_s}
\alias{u_R}
\alias{u_Sv}
\alias{u_acre}
\alias{u_amu}
\alias{u_angstrom}
\alias{u_arc_minute}
\alias{u_arc_second}
\alias{u_astronomical_unit}
\alias{u_atm}
\alias{u_atmosphere}
\alias{u_atomic_mass_unit}
\alias{u_au}
\alias{u_bar}
\alias{u_barn}
\alias{u_becquerel}
\alias{u_btu}
\alias{u_calorie}
\alias{u_cc}
\alias{u_celsius_degree}
\alias{u_centimeter}
\alias{u_cfs}
\alias{u_cm}
\alias{u_cm_per_s}
\alias{u_cubic_centimeter}
\alias{u_cubic_foot}
\alias{u_cubic_inch}
\alias{u_cubic_meter}
\alias{u_cubic_meter_per_s}
\alias{u_cubm_per_s}
\alias{u_cup}
\alias{u_curie}
\alias{u_cycles_per_second}
\alias{u_d}
\alias{u_darcy}
\alias{u_day}
\alias{u_dry_gallon_usa}
\alias{u_dyne}
\alias{u_eV}
\alias{u_electronvolt}
\alias{u_erg}
\alias{u_fahrenheit_degree}
\alias{u_foot}
\alias{u_ft}
\alias{u_g}
\alias{u_gpd}
\alias{u_gpd_per_ft}
\alias{u_gpd_per_sqft}
\alias{u_gpm}
\alias{u_gram}
\alias{u_gramme}
\alias{u_gray}
\alias{u_ha}
\alias{u_hectare}
\alias{u_hectopascal}
\alias{u_horsepower_electric}
\alias{u_horsepower_imperial}
\alias{u_horsepower_metric}
\alias{u_horsepower_water}
\alias{u_hour}
\alias{u_hr}
\alias{u_imperial_gallon_uk}
\alias{u_in}
\alias{u_inH2O}
\alias{u_inHg}
\alias{u_inch}
\alias{u_inchH2O}
\alias{u_inchHg}
\alias{u_kilocalorie}
\alias{u_kilometer}
\alias{u_kilometer_per_hour}
\alias{u_km}
\alias{u_knot}
\alias{u_kph}
\alias{u_lb}
\alias{u_liquid_gallon_usa}
\alias{u_liter}
\alias{u_litre}
\alias{u_mL}
\alias{u_m_per_day}
\alias{u_mb}
\alias{u_mbar}
\alias{u_metric_ton}
\alias{u_mg}
\alias{u_microS_per_cm}
\alias{u_micrometer}
\alias{u_micron}
\alias{u_mil}
\alias{u_mile}
\alias{u_mile_per_hour}
\alias{u_millibar}
\alias{u_milligram}
\alias{u_milliliter}
\alias{u_millimeter}
\alias{u_millimeterHg}
\alias{u_min}
\alias{u_minute}
\alias{u_mm}
\alias{u_mmHg}
\alias{u_molal}
\alias{u_mph}
\alias{u_nautical_mile}
\alias{u_ounce_volume}
\alias{u_parsec}
\alias{u_pc}
\alias{u_per_bar}
\alias{u_pint}
\alias{u_pl}
\alias{u_poise}
\alias{u_poiseuille}
\alias{u_pound_force}
\alias{u_pound_mass}
\alias{u_ppm}
\alias{u_psi}
\alias{u_quart}
\alias{u_rankine}
\alias{u_rem}
\alias{u_roentgen}
\alias{u_rotations_per_minute}
\alias{u_rpm}
\alias{u_sievert}
\alias{u_sk}
\alias{u_slug}
\alias{u_solar_mass}
\alias{u_sqcm}
\alias{u_sqft}
\alias{u_sqft_per_pound}
\alias{u_sqft_per_s}
\alias{u_sqin}
\alias{u_sqkm}
\alias{u_sqm}
\alias{u_sqm_per_N}
\alias{u_sqm_per_day}
\alias{u_sqmi}
\alias{u_sqmm}
\alias{u_square_centimeter}
\alias{u_square_foot}
\alias{u_square_inch}
\alias{u_square_kilometer}
\alias{u_square_meter}
\alias{u_square_mile}
\alias{u_square_millimeter}
\alias{u_square_yard}
\alias{u_sqyd}
\alias{u_statute_mile}
\alias{u_stokes}
\alias{u_tablespoon}
\alias{u_teaspoon}
\alias{u_therm}
\alias{u_ton}
\alias{u_tonne}
\alias{u_torr}
\alias{u_torricelli}
\alias{u_u}
\alias{u_ua}
\alias{u_y}
\alias{u_yard}
\alias{u_yd}
\alias{u_year}
\alias{u_year_sidereal}
\alias{u_year_tropical}
\alias{u_yr}
\title{NON_SI_units}
\format{An object of class \code{physical_quantity} of length 2.}
\usage{
u_arc_minute

u_arc_second

u_angstrom

u_micrometer

u_micron

u_millimeter

u_mm

u_centimeter

u_cm

u_kilometer

u_km

u_nautical_mile

u_astronomical_unit

u_au

u_ua

u_mil

u_inch

u_in

u_foot

u_ft

u_yard

u_yd

u_statute_mile

u_mile

u_parsec

u_pc

u_square_millimeter

u_sqmm

u_square_centimeter

u_sqcm

u_square_meter

u_sqm

u_square_kilometer

u_sqkm

u_barn

u_hectare

u_ha

u_square_inch

u_sqin

u_square_foot

u_sqft

u_square_yard

u_sqyd

u_square_mile

u_sqmi

u_acre

u_cubic_meter

u_liter

u_litre

u_L

u_cubic_centimeter

u_cc

u_milliliter

u_mL

u_cubic_inch

u_cubic_foot

u_imperial_gallon_uk

u_dry_gallon_usa

u_liquid_gallon_usa

u_quart

u_pint

u_cup

u_ounce_volume

u_tablespoon

u_teaspoon

u_minute

u_min

u_hour

u_hr

u_day

u_d

u_year

u_yr

u_y

u_year_sidereal

u_year_tropical

u_kilometer_per_hour

u_kph

u_knot

u_mile_per_hour

u_mph

u_cycles_per_second

u_rotations_per_minute

u_rpm

u_gram

u_gramme

u_g

u_milligram

u_mg

u_tonne

u_metric_ton

u_atomic_mass_unit

u_amu

u_u

u_pound_mass

u_slug

u_ton

u_solar_mass

u_dyne

u_pound_force

u_lb

u_hectopascal

u_bar

u_millibar

u_mbar

u_mb

u_atmosphere

u_atm

u_inchH2O

u_inH2O

u_inchHg

u_inHg

u_millimeterHg

u_mmHg

u_torricelli

u_torr

u_psi

u_electronvolt

u_eV

u_erg

u_btu

u_calorie

u_kilocalorie

u_therm

u_horsepower_electric

u_horsepower_imperial

u_horsepower_metric

u_horsepower_water

u_celsius_degree

u_fahrenheit_degree

u_rankine

u_curie

u_Ci

u_becquerel

u_Bq

u_gray

u_Gy

u_sievert

u_Sv

u_rem

u_roentgen

u_R

u_poiseuille

u_pl

u_poise

u_stokes

u_sk

u_Pa_per_s

u_cubic_meter_per_s

u_cubm_per_s

u_L_per_s

u_gpm

u_cfs

u_m_per_day

u_cm_per_s

u_gpd

u_gpd_per_sqft

u_darcy

u_sqm_per_day

u_sqft_per_s

u_gpd_per_ft

c_volume_solvent

c_volume_solution

c_mass_solvent

c_mass_solution

u_ppm

u_Molar

u_M

u_molal

u_microS_per_cm

u_sqft_per_pound

u_sqm_per_N

u_per_bar
}
\description{
Non-SI units
}
\keyword{datasets}

