% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{physical_quantity}
\alias{physical_quantity}
\title{physical_quantity}
\usage{
physical_quantity(value, units)
}
\arguments{
\item{value}{: The magnitude of the quantity}

\item{units}{: vector to define the specific dimensions of the quantity to create}
}
\value{
A physical quantity with a given value and units
}
\description{
Define a physical quantity
}
\examples{
length <- 1.5*u_m
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

