% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{+.physical_quantity}
\alias{+}
\alias{+.physical_quantity}
\title{+.physical_quantity}
\usage{
a + b
}
\arguments{
\item{a, b}{: these can be a numeric object and/or a physical_quantity object}
}
\value{
A physical quantity
}
\description{
Define the sum operation of physical_quantity class
}
\examples{
dist1 <- 3.4*u_m
dist2 <- 450*u_cm
dist3 <- dist1 + dist2
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

