% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{^.physical_quantity}
\alias{^.physical_quantity}
\title{exponentiation operator}
\usage{
\method{^}{physical_quantity}(a, b)
}
\arguments{
\item{a}{: Physical quantity object}

\item{b}{: It can be a physical quantity object (dimensionless) or a number}
}
\value{
:
The result of this function is a physical quantity
}
\description{
Function to apply the exponentiation operator to physical quantities
}
\examples{
m <- 1.5*u_kg
v <- 3.2*u_m/u_s
v2 <- v^2
energy <- 0.5*m*v2
energy
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

