% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{print.physical_quantity}
\alias{print}
\alias{print.physical_quantity}
\title{print physical_quantity}
\usage{
\\method{print}{physical_quantity}(x, ...)
}
\arguments{
\item{x}{: A physical_quantity}

\item{...}{: Other parameters used in the underlying print function}
}
\value{
Nothing
}
\description{
Function to display a physical_quantity object on the console
}
\examples{
dist <- 1*u_m
print(dist)

}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

