% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{package}
\name{rphysunits}
\alias{rphysunits}
\alias{rphysunits-package}
\title{rphysunits: A Package to define and operate physical quantities in the SI system within R.}
\arguments{
\item{1.0, units}{: value and units}
}
\description{
The rphysunits package includes functions to operate on:
- physical_quantities: A number with an associated set of units. These physical_quantity
  objects are defined using the prefix u_
- physical_constants: A specific constant with its corresponding units. These constants are
  defined using the prefix c_

Basic units
}
\section{rphysunits functions}{

The rphysunits functions are used to define and operate physical_quantities
}

\section{SI Units}{

In this section, a set of objects defining units and constants
}

