% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{same_units}
\alias{same_units}
\title{same_units}
\usage{
same_units(q1, q2)
}
\arguments{
\item{q1, q2}{: Physical_quantity objects}
}
\value{
A logical value
}
\description{
Function to check if two physical_quantity objects have the same units
}
\examples{
a1 <- 5.6*u_m
a2 <- 4.5*u_s
same_units(a1, a2)
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

