% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{set_units}
\alias{set_unit}
\alias{set_units}
\title{set_units}
\usage{
set_units(q, units_label)
}
\arguments{
\item{q}{: physical_quantity}

\item{units_label}{: character}
}
\value{
A physical_quantity
}
\description{
Function to set the units of a physical quantity
}
\examples{
v <- physical_quantity(3.4)
set_units(v, "length")
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

