% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{data}
\name{u_m}
\alias{u_A}
\alias{u_K}
\alias{u_cd}
\alias{u_kg}
\alias{u_m}
\alias{u_mol}
\alias{u_s}
\title{u_m}
\format{An object of class \code{physical_quantity} of length 2.}
\usage{
u_m

u_kg

u_s

u_K

u_A

u_mol

u_cd
}
\description{
Short version of basic SI units
}
\keyword{datasets}

