% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{data}
\name{si_derived_units1}
\alias{si_derived_units1}
\alias{u_coulomb}
\alias{u_degree}
\alias{u_farad}
\alias{u_henry}
\alias{u_hertz}
\alias{u_joule}
\alias{u_lumen}
\alias{u_lux}
\alias{u_newton}
\alias{u_ohm}
\alias{u_pascal}
\alias{u_siemens}
\alias{u_steradian}
\alias{u_tesla}
\alias{u_volt}
\alias{u_watt}
\alias{u_weber}
\title{SI_derived_units}
\format{An object of class \code{physical_quantity} of length 2.}
\usage{
u_steradian

u_degree

u_hertz

u_newton

u_pascal

u_joule

u_watt

u_coulomb

u_volt

u_farad

u_ohm

u_siemens

u_weber

u_tesla

u_henry

u_lumen

u_lux
}
\description{
First set of derived units
}
\keyword{datasets}

