% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{data}
\name{si_derived_units1_short}
\alias{si_derived_units1_short}
\alias{u_C}
\alias{u_F}
\alias{u_Hz}
\alias{u_J}
\alias{u_N}
\alias{u_Pa}
\alias{u_T}
\alias{u_V}
\alias{u_W}
\alias{u_Wb}
\alias{u_deg}
\alias{u_lm}
\alias{u_lx}
\alias{u_sr}
\title{SI_derived_units1_short}
\format{An object of class \code{physical_quantity} of length 2.}
\usage{
u_sr

u_deg

u_Hz

u_N

u_Pa

u_J

u_W

u_C

u_V

u_F

u_Wb

u_T

u_lm

u_lx
}
\description{
First set of derived units (short)
}
\keyword{datasets}

