% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{data}
\name{si_physical_constants}
\alias{c_avogadro}
\alias{c_bohr_magneton}
\alias{c_boltzmann}
\alias{c_e}
\alias{c_electric}
\alias{c_electron_charge}
\alias{c_electron_rest_mass}
\alias{c_elementary_charge}
\alias{c_faraday}
\alias{c_fine_structure}
\alias{c_first_radiation}
\alias{c_g_force}
\alias{c_gravity}
\alias{c_gravity_accel}
\alias{c_h_bar}
\alias{c_ice_point}
\alias{c_magnetic}
\alias{c_molar_gas}
\alias{c_nuclear_magneton}
\alias{c_planck}
\alias{c_proton_rest_mass}
\alias{c_second_radiation}
\alias{c_speed_of_light}
\alias{c_standard_molar_volume}
\alias{c_stefan_boltzmann}
\alias{c_universal_gas}
\alias{c_water_triple_point}
\alias{c_wiens_radiation}
\alias{si_physical_constants}
\title{SI_physical_constants}
\format{An object of class \code{physical_quantity} of length 2.}
\usage{
c_speed_of_light

c_magnetic

c_electric

c_planck

c_h_bar

c_avogadro

c_universal_gas

c_molar_gas

c_standard_molar_volume

c_boltzmann

c_electron_charge

c_elementary_charge

c_e

c_faraday

c_first_radiation

c_second_radiation

c_stefan_boltzmann

c_wiens_radiation

c_electron_rest_mass

c_proton_rest_mass

c_fine_structure

c_bohr_magneton

c_nuclear_magneton

c_gravity

c_gravity_accel

c_g_force

c_ice_point

c_water_triple_point
}
\description{
Physical constants
}
\keyword{datasets}

