% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\docType{data}
\name{si_prefix}
\alias{si_basic_units_short}
\alias{si_prefix}
\alias{u_atto}
\alias{u_centi}
\alias{u_deca}
\alias{u_deci}
\alias{u_deka}
\alias{u_exa}
\alias{u_femto}
\alias{u_giga}
\alias{u_hecto}
\alias{u_kilo}
\alias{u_mega}
\alias{u_micro}
\alias{u_milli}
\alias{u_nano}
\alias{u_peta}
\alias{u_pico}
\alias{u_tera}
\alias{u_yotta}
\alias{u_zepto}
\alias{u_zetta}
\alias{u_zocto}
\title{SI_prefix}
\format{An object of class \code{numeric} of length 1.}
\usage{
u_yotta

u_zetta

u_exa

u_peta

u_tera

u_giga

u_mega

u_kilo

u_hecto

u_deka

u_deca

u_deci

u_centi

u_milli

u_micro

u_nano

u_pico

u_femto

u_atto

u_zepto

u_zocto
}
\description{
Prefixes of the SI units
}
\keyword{datasets}

