% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{/.physical_quantity}
\alias{/.physical_quantity}
\title{/.physical_quantity}
\usage{
\method{/}{physical_quantity}(a, b)
}
\arguments{
\item{a, b}{: these can be a numeric object and/or a physical_quantity object}
}
\value{
a physical_quantity
}
\description{
Define the division operation of physical_quantity class
}
\examples{
vel <- 2.5*u_m/u_s
den <- 2.6*u_gram/u_cubic_centimeter
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

