% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{sqrt.physical_quantity}
\alias{sqrt.physical_quantity}
\title{sqrt}
\usage{
\method{sqrt}{physical_quantity}(x)
}
\arguments{
\item{x}{: Physical quantity object}
}
\value{
the square root of a physical quantity
}
\description{
Square root of a physical quantity
}
\examples{
area <- 1.7*u_m*u_m
length <- sqrt(area)
length
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

