% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{*.physical_quantity}
\alias{*.physical_quantity}
\title{*.physical_quantity}
\usage{
\method{*}{physical_quantity}(a, b)
}
\arguments{
\item{a, b}{: these can be a numeric object and/or a physical_quantity object}
}
\value{
A physical quantity
}
\description{
Define the product operation of physical_quantity class
}
\examples{
distance1 <- 1.5*u_m
var <- u_m*u_s
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

