% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphysunits.R
\name{**.physical_quantity}
\alias{**.physical_quantity}
\title{**.physical_quantity}
\usage{
"**.physical_quantity"(a, b)
}
\arguments{
\item{a, b}{: these can be a numeric object and/or a physical_quantity object}
}
\value{
A physical quantity
}
\description{
Define the power operation of physical_quantity class
}
\examples{
dist <- 1.5*u_m
area <- dist**2
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}

